+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Refactor all files to use Strict Dependency Injection
+
 ## [0.2.0] - 2015-11-18
 ### Added
 - Support for **_Boolean_** type (as checkbox)
@@ -16,5 +20,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.0.1 - 2015-11-14
 
+[Unreleased]: https://github.com/wieldo/meteor-autoformly/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/wieldo/meteor-autoformly/compare/v0.1.0...v0.2.0 
 [0.1.0]: https://github.com/wieldo/meteor-autoformly/compare/v0.0.1...v0.1.0